Refactor db access code
