+++ db/process_message.c
@@ -2475,6 +2475,7 @@ clipper_usage:
         debug_trap(line + st, lline - st);
     }
 
+    // TODO(NC): deprecate?
     /*
        access set password <user>  <password>
        access set read     <table> <user>
