+++ db/toblock.c
@@ -6247,40 +6247,3 @@ int get_next_seqno(void *tran, long long *seqno)
     return 0;
 }
 
-int access_control_check_read(struct ireq *iq, tran_type *trans, int *bdberr)
-{
-    int rc = 0;
-
-    if (gbl_uses_accesscontrol_tableXnode) {
-        rc = bdb_access_tbl_read_by_mach_get(iq->dbenv->bdb_env, trans,
-                                             iq->usedb->tablename,
-                                             nodeix(iq->frommach), bdberr);
-        if (rc <= 0) {
-            reqerrstr(iq, ERR_ACCESS,
-                      "Read access denied to %s from %s bdberr=%d\n",
-                      iq->usedb->tablename, iq->corigin, *bdberr);
-            return ERR_ACCESS;
-        }
-    }
-
-    return 0;
-}
-
-int access_control_check_write(struct ireq *iq, tran_type *trans, int *bdberr)
-{
-    int rc = 0;
-
-    if (gbl_uses_accesscontrol_tableXnode) {
-        rc = bdb_access_tbl_write_by_mach_get(iq->dbenv->bdb_env, trans,
-                                              iq->usedb->tablename,
-                                              nodeix(iq->frommach), bdberr);
-        if (rc <= 0) {
-            reqerrstr(iq, ERR_ACCESS,
-                      "Write access denied to %s from %s bdberr=%d\n",
-                      iq->usedb->tablename, iq->corigin, *bdberr);
-            return ERR_ACCESS;
-        }
-    }
-
-    return 0;
-}
