+++ sqlite/src/comdb2build.h
@@ -3,6 +3,7 @@
 
 #include <sqliteInt.h>
 #include <schemachange.h>
+#include <db_access.h>
 
 #define SQLITE_OPEN_READWRITE        0x00000002  /* Ok for sqlite3_open_v2() */
 #define SQLITE_OPEN_CREATE           0x00000004  /* Ok for sqlite3_open_v2() */
